refactor: move addMulti to numbers
